requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via packaging, pytest
configparser==3.8.1       # via importlib-metadata
contextlib2==0.5.5        # via importlib-metadata
coverage==4.5.4           # via pytest-cov
ddt==1.2.1
django-model-utils==3.2.0
django-ratelimit-backend==1.1.1
django-ratelimit==2.0.0
factory-boy==2.8.1
faker==2.0.1              # via factory-boy
funcsigs==1.0.2           # via mock, pytest
importlib-metadata==0.19  # via pluggy, pytest
ipaddress==1.0.22         # via faker
mock==3.0.5
more-itertools==5.0.0     # via pytest, zipp
packaging==19.1           # via pytest
pathlib2==2.3.4           # via importlib-metadata, pytest, pytest-django
pluggy==0.12.0            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.2          # via packaging
pytest-cov==2.7.1
pytest-django==3.5.1
pytest==4.6.5             # via pytest-cov, pytest-django
python-dateutil==2.4.0
pytz==2019.2              # via django
scandir==1.10.0           # via pathlib2
six==1.12.0               # via faker, mock, more-itertools, packaging, pathlib2, pytest, python-dateutil
text-unidecode==1.2       # via faker
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata
